fix(text): compute the SDF edge in fp32
The text SDF and MSDF fragment stages declared `mediump`. Their antialiased
band is a few thousandths of the distance field wide, while fp16's spacing
around the 0.5 threshold is 2^-11 — so a driver that honours `mediump` as real
half-float quantises the whole ramp to about ten distinct intensities and the
edge reads as a hard, blocky step.

Desktop ANGLE and SwiftShader compute `mediump` in fp32 and never showed it;
Mesa's llvmpipe, which the Firefox CI lane runs on, does not: a magnified glyph
measured 10 distinct intensities there against more than 200 elsewhere.

+++ src/rendering/webgl2/glsl/text-sdf.frag
@@ -1,5 +1,11 @@
 #version 300 es
-precision mediump float;
+// The edge math runs in fp32 even though nothing else here needs it: the
+// antialiased band is a few thousandths of the field wide, while fp16's spacing
+// around the 0.5 threshold is 2^-11. A driver that honours `mediump` as real
+// half-float (Mesa's llvmpipe does; desktop ANGLE and SwiftShader do not)
+// therefore quantises the whole ramp to ~10 distinct intensities, which reads
+// as a hard, blocky edge on a magnified glyph.
+precision highp float;
 
 uniform sampler2D u_nodeData;  // RGBA32F per-node data (see WebGl2TextRenderer)
 uniform float     u_pageSize;  // atlas page size in px (for shadow UV conversion)
